refactored structure of project
