+++ src/task1/task1.js
@@ -1,0 +1,8 @@
+const getRandomNumber = () => {
+    console.log('getRandomNumber() started');
+    const result = Math.random();
+    console.log('getRandomNumber() result = ', result);
+    return result;
+}
+
+getRandomNumber();
