+++ src/task3/EventEmitter.js
