+++ src/task3/task2.js
